This commit is conceptually orthogonal to the other pull request ("IndependentSolverGetInitFactorizer") I submitted today (2/10). Therefore, I am submitting them as two different patches.  In actuality, however, they play off of each other for a combined effect superior to either one individually.

This particular commit targets the CexCachingSolver so that the UBTree is called less frequently.  Specifically, the two optimizations are a "QuickCache" that determines when two constraints are identical.  The cache stores solutions so that they can be used in the future.  The second optimizations tests the parent state's solution on every newly created state.  This will work, by defn, at least
50% of the time.  It also prevents the UBTree's invocation, which can be costly.

This pull request's improvements are most pronounced when combined with commit the factorization of the getInitialValues that I did in my other pull request.  The time spent in the state dump stage is significantly reduced, often resulting in 100% improvements when executing the default search.

After completion, the changes:
1) Pass all provided tests (other than ones that were expected to fail)
2) Have been verified via -debug-validate-solver by a complete run through
   of 9 members of the bin-utils for
     --search=dfs with a halt being placed, and states dumped, after 500 forks
     --search=default with a halt being placed, and the states dumped, after 1000 forks
3) Achieved significant speedup on test programs
     --search=dfs
           a. On average 93% improvement (for example, nm-new saw a decrease from 1200 seconds to 670 seconds)
           b. Worst case 13% improvement (obj-copy when from 950 seconds to 835 seconds)
     --search=default
           a. On average 85% improvement (for example, addr2line saw a decrease from 1325 seconds to 717 seconds)
           b. Worst case 29% improvement (ar went from 560 seconds to 435 seconds)